+++ client/main.html
@@ -49,11 +49,6 @@
     {{> addPool}}
 </template>
 
-<template name="appPoolAppend">
-     <!--@TODO appPoolEdit-->
-    {{> appendPool}}
-</template>
-
 <template name="appHistory">
     <!--@TODO appHistory-->
     {{> history}}
