+++ client/main.js
@@ -16,8 +16,6 @@ import './../imports/ui/components/common/common.js';
 
 import './../imports/ui/components/add_pool/add_pool.html';
 import './../imports/ui/components/add_pool/add_pool.js';
-import './../imports/ui/components/append_pool/append_pool.html';
-import './../imports/ui/components/append_pool/append_pool.js';
 
 import './../imports/ui/components/pools_list/pools_list.html';
 import './../imports/ui/components/pools_list/pools_list.js';
