+++ client/main.styl
@@ -6,6 +6,9 @@
 @import '../imports/ui/components/feeds/feeds.styl'
 @import '../imports/ui/components/auth/auth.styl'
 @import '../imports/ui/components/pools_list/pools_list.styl'
+@import '../imports/ui/components/pool_pending/pool_pending.styl'
+@import '../imports/ui/components/pool_summary/pool_summary.styl'
+@import '../imports/ui/components/pool_archived/pool_archived.styl'
 @import '../imports/ui/components/add_pool/add_pool.styl'
 @import '../imports/ui/components/notifications/notifications.styl'
 @import '../imports/ui/components/pools_list_item/pools_list_item.styl'
