--- imports/ui/components/append_pool/append_pool.html
@@ -1,28 +1,0 @@
-<template name="appendPool">
-    {{>poolPending}}
-    <div class="row content-container">
-        <div class="container">
-            <div class="list-group col-xs-12 col-sm-8 col-sm-offset-2">
-                <form id="append_pool">
-                    <div class="form-group new-product-input-group">
-                        <label class="control-label" for="products">Ссылка на продукт</label>
-                        <div class="controls">
-                            <input class="form-control" name="item-link" id="new-product-input">
-                        </div>
-                    </div>
-                    {{#each itemFields}}
-                        <div class="form-group new-product-input-group">
-                            <label class="control-label" for="products">{{title}}</label>
-                            <div class="controls">
-                                <input name={{key}} class="form-control" value={{value}} id="new-product-input">
-                            </div>
-                        </div>
-                    {{/each}}
-                    <div class="input-group">
-                        <button class="btn btn-success join-button disabled" data-loading-text="<i class='fa fa-spinner fa-spin '></i> Загружаем">Присоединиться</button>
-                    </div>
-                </form>
-            </div>
-        </div>
-    </div>
-</template>
