--- imports/ui/components/append_pool/append_pool.styl
@@ -1,1 +1,0 @@
-// append_pool.styl
