+++ imports/ui/components/common/common.styl
@@ -6,3 +6,12 @@ body
 
 .content-container
   padding-top 80px
+
+.clearfix:before,
+.clearfix:after
+  content " "
+  display table
+.clearfix:after
+  clear both
+.clearfix
+  *zoom 1
