+++ lib/router.js
@@ -72,21 +72,6 @@ Router.route('/pool/add', function() {
     }
 }, { name: 'pool_add' });
 
-Router.route('/pool/:poolId/append', function() {
-    this.wait(Meteor.subscribe('PoolsOne', this.params.poolId));
-    this.wait(Meteor.subscribe('PoolsOrders', this.params.poolId));
-    this.wait(Meteor.subscribe('OrdersItems', this.params.poolId));
-    if (this.ready()) {
-        this.render('appPoolAppend', {
-            data: () => {
-                return Pools.findOne({ _id: this.params.poolId });
-            }
-        });
-    }else {
-        this.render('loading');
-    }
-}, { name: 'pool_append' });
-
 Router.route('/pool/:poolId', function() {
     this.wait(Meteor.subscribe('PoolsOne', this.params.poolId));
     this.wait(Meteor.subscribe('PoolsOrders', this.params.poolId));
